app/requirements.txt
mysql-connector-python
Flask
PyJWT 
prometheus_client